+++ src/OpenTelemetry/Metrics/MetricReaderType.cs
@@ -1,4 +1,4 @@
-// <copyright file="ExplicitBucketHistogramConfiguration.cs" company="OpenTelemetry Authors">
+// <copyright file="MetricReaderType.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,15 +16,21 @@
 
 namespace OpenTelemetry.Metrics
 {
-    public class ExplicitBucketHistogramConfiguration : MetricStreamConfiguration
+    /// <summary>
+    /// Type of <see cref="MetricReader" /> to be used.
+    /// </summary>
+    public enum MetricReaderType
     {
         /// <summary>
-        /// Gets or sets the values representing explicit histogram bucket
-        /// boundary values.
+        /// Use the <see cref="BaseExportingMetricReader" />.
+        /// This requires manually invoking <c>MetricReader.Collect()</c> to export metrics.
         /// </summary>
-        /// <remarks>
-        /// The array must be in ascending order with distinct values.
-        /// </remarks>
-        public double[] Boundaries { get; set; }
+        Manual,
+
+        /// <summary>
+        /// Use the <see cref="PeriodicExportingMetricReader" />.
+        /// <c>MetricReader.Collect()</c> will be invoked on a defined interval.
+        /// </summary>
+        Periodic,
     }
 }